pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "securAIty"
version = "0.1.0"
description = "AI-powered security orchestration platform with NATS messaging, PostgreSQL storage, and Vault secrets management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "securAIty Team" }
]
keywords = ["security", "ai", "nats", "postgresql", "vault", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "nats-py>=2.4.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "httpx>=0.26.0",
    "cryptography>=42.0.0",
    "python-jose[cryptography]>=3.3.0",
    "structlog>=24.1.0",
]

[project.urls]
Homepage = "https://github.com/securAIty/securAIty"
Documentation = "https://securAIty.readthedocs.io"
Repository = "https://github.com/securAIty/securAIty"
Issues = "https://github.com/securAIty/securAIty/issues"

[project.scripts]
securAIty = "securAIty.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "bandit[toml]>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/securAIty"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/pyproject.toml",
    "/README.md",
    "/LICENSE",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "e2e: marks end-to-end tests",
    "unit: marks unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
implicit_reexport = false
strict_equality = true
extra_checks = true
show_error_codes = true
show_traceback = true
pretty = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "nats.*",
    "asyncpg.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ruff]
line-length = 88
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "N",
    "SIM",
    "PTH",
    "RUF",
]
ignore = [
    "E501",
    "D100",
    "D104",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG", "PLR2004"]
"src/securAIty/api/**/*.py" = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["securAIty"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.bandit]
targets = ["src/securAIty"]
exclude_dirs = ["tests"]
skips = ["B101", "B311"]

[tool.coverage.run]
source = ["src/securAIty"]
branch = true
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/cli.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
